+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 * [PQA-42] obtain info from workspace 
 * [SPK-988] Fix runInAllNodes spec
 * [QATM-3515] Improve getHostIp method to obtain different IP's if service has multiple instances
+* [SAAS-2785] New step to get IP (internal and external) from Marathon instead of CCT
 
 ## 0.19.1 (September 16, 2020)